Fix a race condition in boot-driver (#379)
get-free-port finds a free port, but does not hold on to it. etaoin
later uses this port, but sometimes the port is no longer available and
at that point etaoin throws an exception.

This is fixed by catching the exception and retrying with a new port
choice.

I split out the driver and opts creation into a `boot-driver*` fn,
because otherwise `boot-driver` would have become too complex.